+++ src/sync/mod.rs
@@ -1,0 +1,5 @@
+// Copyright 2018-2024 the Deno authors. MIT license.
+
+mod flag;
+
+pub use flag::AtomicFlag;
