+++ src/tokio/mod.rs
@@ -1,0 +1,15 @@
+// Copyright 2018-2024 the Deno authors. MIT license.
+
+pub use joinset::JoinSet;
+pub use split::split_io;
+pub use split::IOReadHalf;
+pub use split::IOWriteHalf;
+
+pub use task::spawn;
+pub use task::spawn_blocking;
+pub use task::JoinHandle;
+pub use task::MaskFutureAsSend;
+
+mod joinset;
+mod split;
+mod task;
